% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocini}
\alias{ocini}
\title{store owncloud credentials}
\usage{
ocini(url, user, pwd)
}
\arguments{
\item{url}{owncloud url}

\item{user}{user-name (often this is the user email)}

\item{pwd}{user-password}
}
\description{
store owncloud credentials
}
\author{
valcu@orn.mpg.de, \email{valcu@orn.mpg.de}
}
