% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc.R
\name{owncloud}
\alias{owncloud}
\title{owncloud}
\usage{
owncloud(dir, basedir = "~/ownCloud/", exclude = c("*.sublime-workspace"))
}
\arguments{
\item{dir}{oc directory to synchronize, relative to basedir.
dir can be missing only if the root directory was synchronized already.}

\item{basedir}{local directory. default to ~/ownCloud/}

\item{exclude}{files and/or file types to exclude. default to "*.sublime-workspace"}
}
\value{
owncloudcmd verbose output returned invisibly.
}
\description{
An owncloud client adapted for R.
}
