% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocini}
\alias{ocini}
\title{store owncloud credentials}
\usage{
ocini(url, user, pwd)
}
\arguments{
\item{user}{user-name (often this is the user email)}

\item{pwd}{user-password}
}
\description{
The password is stored encrypted using 
\href{http://manpages.ubuntu.com/manpages/bionic/man1/secret-tool.1.html}{secret-tool}.
}
\author{
valcu@orn.mpg.de, \email{valcu@orn.mpg.de}
}
